+++ examples/carbon-for-ibm-products/gallery-examples.test.js
@@ -11,6 +11,7 @@ import { init } from './test-common';
 
 import { Example as APIKeyModalExample } from './APIKeyModal/src/Example/Example';
 import { Example as AboutModalExample } from './AboutModal/src/Example/Example';
+import { Example as BigNumbersExample } from './BigNumbers/src/Example/Example';
 import { Example as CascadeExample } from './Cascade/src/Example/Example';
 import { Example as CoachmarkExample } from './Coachmark/src/Example/Example';
 import { Example as CoachmarkBeaconExample } from './CoachmarkBeacon/src/Example/Example';
@@ -31,6 +32,7 @@ import { Example as EditInPlaceExample } from './EditInPlace/src/Example/Example
 import { Example as EmptyStatesExample } from './EmptyStates/src/Example/Example';
 import { Example as ExportModalExample } from './ExportModal/src/Example/Example';
 import { Example as ExpressiveCardExample } from './ExpressiveCard/src/Example/Example';
+import { Example as FullPageErrorExample } from './FullPageError/src/Example/Example';
 import { Example as HTTPErrorsExample } from './HTTPErrors/src/Example/Example';
 import { Example as ImportModalExample } from './ImportModal/src/Example/Example';
 import { Example as InterstitialScreenExample } from './InterstitialScreen/src/Example/Example';
@@ -45,9 +47,11 @@ import { Example as SavingExample } from './Saving/src/Example/Example';
 import { Example as SearchBarExample } from './SearchBar/src/Example/Example';
 import { Example as SidePanelExample } from './SidePanel/src/Example/Example';
 import { Example as StatusIconExample } from './StatusIcon/src/Example/Example';
+import { Example as StringFormatterExample } from './StringFormatter/src/Example/Example';
 import { Example as TagSetExample } from './TagSet/src/Example/Example';
 import { Example as TearsheetExample } from './Tearsheet/src/Example/Example';
 import { Example as TruncatedListExample } from './TruncatedList/src/Example/Example';
+import { Example as UserAvatarExample } from './UserAvatar/src/Example/Example';
 import { Example as UserProfileImageExample } from './UserProfileImage/src/Example/Example';
 import { Example as WebTerminalExample } from './WebTerminal/src/Example/Example';
 import { Example as PrefixExampleExample } from './prefix-example/src/Example/Example';
@@ -65,6 +69,12 @@ describe('All examples', () => {
 
   it('AboutModal renders', () => {
     render(<AboutModalExample />);
+    // expect no errors int the console
+    expect(console.error).not.toHaveBeenCalled();
+  });
+
+  it('BigNumbers renders', () => {
+    render(<BigNumbersExample />);
     // expect no errors int the console
     expect(console.error).not.toHaveBeenCalled();
   });
@@ -189,6 +199,12 @@ describe('All examples', () => {
     expect(console.error).not.toHaveBeenCalled();
   });
 
+  it('FullPageError renders', () => {
+    render(<FullPageErrorExample />);
+    // expect no errors int the console
+    expect(console.error).not.toHaveBeenCalled();
+  });
+
   it('HTTPErrors renders', () => {
     render(<HTTPErrorsExample />);
     // expect no errors int the console
@@ -273,6 +289,12 @@ describe('All examples', () => {
     expect(console.error).not.toHaveBeenCalled();
   });
 
+  it('StringFormatter renders', () => {
+    render(<StringFormatterExample />);
+    // expect no errors int the console
+    expect(console.error).not.toHaveBeenCalled();
+  });
+
   it('TagSet renders', () => {
     render(<TagSetExample />);
     // expect no errors int the console
@@ -287,6 +309,12 @@ describe('All examples', () => {
 
   it('TruncatedList renders', () => {
     render(<TruncatedListExample />);
+    // expect no errors int the console
+    expect(console.error).not.toHaveBeenCalled();
+  });
+
+  it('UserAvatar renders', () => {
+    render(<UserAvatarExample />);
     // expect no errors int the console
     expect(console.error).not.toHaveBeenCalled();
   });