+++ packages/ibm-products/src/components/index.js
@@ -20,6 +20,7 @@ export {
 export { CreateTearsheetNarrow } from './CreateTearsheetNarrow';
 export {
   EmptyState,
+  EmptyStateV2,
   ErrorEmptyState,
   NoDataEmptyState,
   NoTagsEmptyState,
@@ -97,9 +98,13 @@ export {
 } from './Guidebanner';
 export { InlineTip, InlineTipButton, InlineTipLink } from './InlineTip';
 export { NonLinearReading } from './NonLinearReading';
+export { BigNumbers } from './BigNumbers';
 export { TruncatedList } from './TruncatedList';
 export { InterstitialScreen } from './InterstitialScreen';
 export { InterstitialScreenView } from './InterstitialScreenView';
 export { InterstitialScreenViewModule } from './InterstitialScreenViewModule';
 export { DelimitedList } from './DelimitedList';
+export { FullPageError } from './FullPageError';
 export { SearchBar } from './SearchBar';
+export { StringFormatter } from './StringFormatter';
+export { UserAvatar } from './UserAvatar';